+++ app/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -1425,7 +1425,8 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
             snapMessages,
             ethScheduler,
             protocolContext,
-            synchronizer));
+            synchronizer,
+            metricsSystem));
   }
 
   WorldStateArchive createWorldStateArchive(